Introduit la classe Restitution::AvecEntrainement pour alléger la classe Restitution::Securite
