+++ app/models/restitution/base.rb
@@ -87,6 +87,10 @@ module Restitution
       self.class.const_get('COMPETENCES_MOBILISEES')
     end
 
+    def premier_evenement_du_nom(evenements, nom_evenement)
+      evenements.find { |e| e.nom == nom_evenement }
+    end
+
     private
 
     def competences_indeterminees?(competences)
